+++ src/Filesystem/TwigTemplateLocator.php
@@ -9,11 +9,16 @@
 namespace HeimrichHannot\TwigSupportBundle\Filesystem;
 
 use Contao\CoreBundle\Config\ResourceFinderInterface;
+use Contao\CoreBundle\Routing\ScopeMatcher;
+use Contao\PageModel;
+use Contao\ThemeModel;
+use Contao\Validator;
 use HeimrichHannot\TwigSupportBundle\Cache\TemplateCache;
 use HeimrichHannot\TwigSupportBundle\Exception\TemplateNotFoundException;
-use SplFileInfo;
 use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Finder\Finder;
+use Symfony\Component\Finder\SplFileInfo;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Webmozart\PathUtil\Path;
 
@@ -28,26 +33,188 @@ class TwigTemplateLocator
      */
     protected $contaoResourceFinder;
 
-    public function __construct(KernelInterface $kernel, ResourceFinderInterface $contaoResourceFinder)
+    protected RequestStack $requestStack;
+
+    protected ScopeMatcher $scopeMatcher;
+
+    public function __construct(KernelInterface $kernel, ResourceFinderInterface $contaoResourceFinder, RequestStack $requestStack, ScopeMatcher $scopeMatcher)
     {
         $this->kernel = $kernel;
         $this->contaoResourceFinder = $contaoResourceFinder;
+        $this->requestStack = $requestStack;
+        $this->scopeMatcher = $scopeMatcher;
     }
 
     /**
      * Return a twig template path by template name (without or with extension).
      *
+     * Options:
+     * - (bool) disableCache: Set to true to disable cache. Cache is disabled by default in dev environment. Default false.
+     *
      * @throws TemplateNotFoundException
      */
-    public function getTemplatePath(string $name, bool $disableCache = false): string
+    public function getTemplatePath(string $templateName, array $options = []): string
     {
-        if (($templates = $this->getTemplates(false, $disableCache)) && \array_key_exists($name, $templates)) {
-            return $templates[$name];
-        } elseif (($templates = $this->getTemplates(true, $disableCache)) && \array_key_exists($name, $templates)) {
-            return $templates[$name];
+        $templateName = basename($templateName);
+        $themeFolder = '';
+        $disableCache = isset($configuration['disableCache']) && true === $configuration['disableCache'];
+
+        if ($this->scopeMatcher->isFrontendRequest($this->requestStack->getCurrentRequest())) {
+            /* @var PageModel $objPage */
+            global $objPage;
+
+            if ('' != $objPage->templateGroup) {
+                if (Validator::isInsecurePath($objPage->templateGroup)) {
+                    throw new \RuntimeException('Invalid path '.$objPage->templateGroup);
+                }
+
+                $themeFolder = $objPage->templateGroup;
+            }
         }
 
-        throw new TemplateNotFoundException(sprintf('Unable to find template "%s".', $name));
+        if ((!$templates = $this->getTemplates(false, $disableCache)) || !\array_key_exists($templateName, $templates)) {
+            if ((!$templates = $this->getTemplates(true, $disableCache)) || !\array_key_exists($templateName, $templates)) {
+                throw new TemplateNotFoundException(sprintf('Unable to find template "%s".', $templateName));
+            }
+        }
+
+        $template = $templates[$templateName];
+
+        if ('templates' === substr($themeFolder, 0, 9)) {
+            $themeFolder = trim(substr($themeFolder, 9), '/');
+        }
+        $pathLength = \strlen($themeFolder);
+
+        foreach ($template['paths'] as $path) {
+            if ($themeFolder === substr($path, 0, $pathLength)) {
+                return $path;
+            }
+        }
+
+        foreach ($template['paths'] as $path) {
+            if ('@' !== substr($path, 0, 1)) {
+                return $path;
+            }
+        }
+
+        return end($template['paths']);
+    }
+
+    /**
+     * Return all twig template files of a particular group as array.
+     *
+     * Configuration options:
+     * - (bool) extension: Return the file extension in the file names. Default false.
+     * - (bool) disableCache: Set to true to disable cache. Cache is disabled by default in dev environment. Default false.
+     *
+     * , array $arrAdditionalMapper=array(), $strDefaultTemplate='', string $fileExtension
+     *
+     * @param string|string[] $prefixes
+     * @param array           $arrAdditionalMapper
+     * @param string          $strDefaultTemplate
+     * @param string          $fileExtension
+     */
+    public function getTemplateGroup($prefixes, array $configuration = [])
+    {
+        if (!\is_array($prefixes) && !\is_string($prefixes)) {
+            throw new \InvalidArgumentException('Only string or array are allowed!');
+        }
+
+        if (\is_string($prefixes)) {
+            $prefixes = [$prefixes];
+        }
+
+        $templateNames = [];
+        $disableCache = isset($configuration['disableCache']) && true === $configuration['disableCache'];
+        $extension = isset($configuration['extension']) && true === $configuration['extension'];
+
+        foreach ($prefixes as $prefix) {
+            $templateNames = array_merge($templateNames,
+                $this->getPrefixedFiles($prefix,
+                    $extension,
+                    $disableCache)
+            );
+        }
+
+        try {
+            $objTheme = ThemeModel::findAll(['order' => 'name']);
+        } catch (\Exception $e) {
+            $objTheme = null;
+        }
+
+        $options = [];
+
+        foreach ($templateNames as $templateName) {
+            if ((!$templates = $this->getTemplates(false, $disableCache)) || !\array_key_exists($templateName, $templates)) {
+                continue;
+            }
+
+            $template = $templates[$templateName];
+
+            $templatePathList = [];
+
+            foreach ($template['paths'] as $path) {
+                if ($path && '@' === substr($path, 0, 1)) {
+                    $templatePathList['bundles'][] = explode('/', $path)[0];
+                } else {
+                    $folders = explode('/', $path);
+
+                    if (\count($folders) <= 1) {
+                        $templatePathList['global'] = $path;
+                    } else {
+                        foreach ($objTheme as $theme) {
+                            $themeFolder = substr($objTheme->templates, 10);
+
+                            if ($themeFolder === substr($path, 0, \strlen($themeFolder))) {
+                                $templatePathList['themefolders'][] = $theme->name;
+                            }
+                        }
+                    }
+                }
+            }
+            $optionLabel = '';
+
+            if (isset($templatePathList['global'])) {
+                $optionLabel .= $GLOBALS['TL_LANG']['MSC']['global'].', ';
+            }
+
+            if (isset($templatePathList['themefolders']) && !empty($templatePathList['themefolders'])) {
+                $optionLabel .= implode(', ', $templatePathList['themefolders']).', ';
+            }
+
+            if (isset($templatePathList['bundles']) && !empty($templatePathList['bundles'])) {
+                $optionLabel .= implode(', ', $templatePathList['bundles']);
+            }
+
+            $options[$templateName] = $templateName.' ('.trim($optionLabel, ', ').')';
+        }
+
+        return $options;
+    }
+
+    /**
+     * Return the files matching a prefix as array.
+     *
+     * Configuration options:
+     * - (bool) extension: Return the file extension in the file names. Default false.
+     * - (bool) disableCache: Set to true to disable cache. Cache is disabled by default in dev environment. Default false.
+     *
+     * @param string $prefix The prefix (e.g. "moo_")
+     *
+     * @throws \Psr\SimpleCache\InvalidArgumentException
+     *
+     * @return array An array of matching files
+     */
+    public function getPrefixedFiles(string $prefix, array $configuration = [])
+    {
+        $disableCache = isset($configuration['disableCache']) && true === $configuration['disableCache'];
+        $extension = isset($configuration['extension']) && true === $configuration['extension'];
+
+        if ('_' != substr($prefix, -1)) {
+            $prefix .= '($|_)';
+        }
+
+        return array_values(preg_grep('/^'.$prefix.'/', array_keys($this->getTemplates($extension, $disableCache))));
     }
 
     /**
@@ -68,17 +235,19 @@ class TwigTemplateLocator
             $cacheItem = $cache->getItem($cacheKey);
 
             if (!$cacheItem->isHit()) {
-                $cacheItem->set($this->getTwigTemplatePaths(false));
+                $cacheItem->set($this->generateContaoTwigTemplatePaths(false));
                 $cache->save($cacheItem);
             }
 
             return $cache->getItem($cacheKey)->get();
         }
 
-        return $this->getTwigTemplatePaths(false);
+        return $this->generateContaoTwigTemplatePaths(false);
     }
 
     /**
+     * Return twig templates in a given path.
+     *
      * @param iterable|string $dir
      */
     public function getTwigTemplatesInPath($dir, ?string $twigKey = null, bool $extension = false): array
@@ -102,9 +271,9 @@ class TwigTemplateLocator
             }
 
             if (!$twigKey) {
-                $twigFiles[$name] = Path::makeRelative($file->getRealPath(), $this->kernel->getProjectDir().'/templates');
+                $twigFiles[$name]['paths'][] = Path::makeRelative($file->getRealPath(), $this->kernel->getProjectDir().'/templates');
             } else {
-                $twigFiles[$name] = "@$twigKey/".$file->getRelativePathname();
+                $twigFiles[$name]['paths'][] = "@$twigKey/".$file->getRelativePathname();
             }
         }
 
@@ -112,11 +281,11 @@ class TwigTemplateLocator
     }
 
     /**
-     * Return all twig file path.
+     * Return all twig file paths in the typical folders.
      *
      * @return array
      */
-    protected function getTwigTemplatePaths(bool $extension = false)
+    protected function generateContaoTwigTemplatePaths(bool $extension = false)
     {
         $bundles = $this->kernel->getBundles();
         $twigFiles = [];
@@ -132,16 +301,16 @@ class TwigTemplateLocator
                 }
                 $twigKey = preg_replace('/Bundle$/', '', $key);
 
-                $twigFiles = array_merge($twigFiles, $this->getTwigTemplatesInPath($dir, $twigKey, $extension));
+                $twigFiles = array_merge_recursive($twigFiles, $this->getTwigTemplatesInPath($dir, $twigKey, $extension));
             }
         }
 
         // Bundle template folders
-        $twigFiles = array_merge($twigFiles, $this->getTwigTemplatesInPath(
+        $twigFiles = array_merge_recursive($twigFiles, $this->getTwigTemplatesInPath(
             $this->contaoResourceFinder->findIn('templates')->name('*.twig')->getIterator(), null, $extension));
 
         // Project template folder
-        $twigFiles = array_merge($twigFiles, $this->getTwigTemplatesInPath(
+        $twigFiles = array_merge_recursive($twigFiles, $this->getTwigTemplatesInPath(
             $this->kernel->getProjectDir().'/templates', null, $extension));
 
         return $twigFiles;